+++ src/main/java/com/kritik/POS/restaurant/models/response/StockReceiptResponse.java
@@ -15,7 +15,7 @@ public class StockReceiptResponse {
     private String invoiceNumber;
     private LocalDateTime receivedAt;
     private Integer totalItems;
-    private Integer totalQuantity;
+    private Double totalQuantity;
     private Double totalCost;
     private String notes;
     private LocalDateTime createdAt;
